move current demos into ./v0_0_2/
